refactor(cli): align configuration with imitation pattern using hydra-zen stores
