--- src/config/cli/factories/__init__.py
@@ -1,11 +1,0 @@
-"""
-Hydra-zen builders for CLI configuration components.
-
-This module exports factory functions that create Hydra-compatible configurations
-from the Pydantic models. Each builder wraps a schema model with the necessary
-Hydra-zen metadata for proper instantiation and configuration management.
-
-The builders follow a consistent naming pattern (build_*) and use zen() wrappers
-to maintain Pydantic validation within the Hydra configuration system.
-"""
-from .cli import *
