+++ src/config/cli/schemas.py
@@ -1,10 +1,21 @@
 """
-User interface and display configuration schemas.
+Configuration schemas for CLI components.
 
-This module defines configuration for the Rich terminal interface including
-themes, styles, messages, and display components used by the ThermurUI helper.
+This module provides all configuration models for the Thermur CLI system. These schemas
+control user interaction, display formatting, system validation, and data management.
+Unlike the training configuration, these schemas are instantiated directly rather than
+through Hydra's runtime, as they configure the CLI framework itself.
+
+The schemas are organized alphabetically and follow consistent patterns for validation
+and defaults. They are consumed by CLI helpers (ThermurUI, CLIPrompts, SystemInspector,
+GlobusManager) to provide a rich terminal experience.
 """
-from pydantic import BaseModel, Field, PositiveInt
+from pathlib           import Path
+from pydantic          import BaseModel, computed_field, Field, PositiveInt, SecretStr
+from pydantic_settings import BaseSettings, SettingsConfigDict
+from typing            import Literal, Optional
+
+SECRETS_DIR = Path.home() / ".config" / "thermur" / "secrets"
 
 
 class DisplayModel(BaseModel, extra="forbid"):
@@ -107,25 +118,6 @@ class DisplayModel(BaseModel, extra="forbid"):
         ],
         description = "Color gradient for fire effects."
     )
-    styles: dict[str, str] = Field(
-        default = {
-            'thermal'   : 'bold orange_red1',
-            'flock'     : 'bold bright_cyan', 
-            'drone'     : 'yellow',
-            'success'   : 'bold green',
-            'warning'   : 'bold yellow',
-            'error'     : 'bold red',
-            'info'      : 'cyan',
-            'accent'    : 'bright_magenta',
-            'muted'     : 'grey50',
-            'bright'    : 'bright_white',
-            'dim'       : 'grey30',
-            'highlight' : 'on dark_blue',
-            'config'    : 'bright_blue',
-            'debug'     : 'italic grey70',
-        },
-        description = "Named styles for consistent theming."
-    )
     message_types: dict[str, dict[str, str]] = Field(
         default = {
             'info'    : {'icon': 'ℹ️ ', 'style': 'info'},
@@ -154,6 +146,25 @@ class DisplayModel(BaseModel, extra="forbid"):
     resource_details_template: str = Field(
         default     = "{:.1f}{} / {:.1f}{}",
         description = "Format template for resource usage display."
+    )
+    styles: dict[str, str] = Field(
+        default = {
+            'thermal'   : 'bold orange_red1',
+            'flock'     : 'bold bright_cyan', 
+            'drone'     : 'yellow',
+            'success'   : 'bold green',
+            'warning'   : 'bold yellow',
+            'error'     : 'bold red',
+            'info'      : 'cyan',
+            'accent'    : 'bright_magenta',
+            'muted'     : 'grey50',
+            'bright'    : 'bright_white',
+            'dim'       : 'grey30',
+            'highlight' : 'on dark_blue',
+            'config'    : 'bright_blue',
+            'debug'     : 'italic grey70',
+        },
+        description = "Named styles for consistent theming."
     )
     system_components: dict[str, str] = Field(
         default = {
@@ -229,6 +240,118 @@ class DisplayModel(BaseModel, extra="forbid"):
     )
 
 
+class DownloadModel(BaseModel, extra="forbid"):
+    """
+    Download and management configuration for the CLI.
+    
+    This model contains settings for file downloads, display options, and
+    caching behavior, separate from the dataset schema used for training.
+    """
+    globus_client_id: str = Field(
+        default     = "ac349f52-8197-4a41-8d6d-5ae1c879273f",
+        description = "Native app client ID for Globus OAuth2 authentication flow."
+    )
+    globus_dataset_path: str = Field(
+        default     = "/1/published/publication_309/submitted_data",
+        description = "Path to the WRF-Fire dataset within the FRDR Globus endpoint."
+    )
+    globus_endpoint_id: str = Field(
+        default     = "f163c1b3-9c88-42f6-a7bb-5839ed6c4063",
+        description = "UUID of the FRDR Globus endpoint hosting WRF-Fire simulations."
+    )
+    globus_scopes: str = Field(
+        default     = "urn:globus:auth:scope:transfer.api.globus.org:all",
+        description = "OAuth2 scopes required for Globus transfer operations."
+    )
+    recommended_files: list[dict[str, str]] = Field(
+        default = [
+            {
+                "file" : "wrfout_W3F1R0",
+                "desc" : "Light wind (3m/s) over short grass. Represents prescribed burns "
+                         "or early-season fires in grasslands with stable atmospheric conditions."
+            },
+            {
+                "file" : "wrfout_W5F7R4",  
+                "desc" : "Moderate wind (5m/s) through brushy forest understory. Models typical "
+                         "wildfire conditions with mixed vegetation and moderate atmospheric mixing."
+            },
+            {
+                "file" : "wrfout_W8F13R6",
+                "desc" : "Strong wind (8m/s) through heavy dead trees and branches. Simulates "
+                         "post-logging or storm damage areas with deep atmospheric mixing."
+            },
+            {
+                "file" : "wrfout_W12F4R8",
+                "desc" : "Extreme wind (12m/s) in dense shrubland. Represents high-risk fire "
+                         "weather in Mediterranean climates with strong temperature inversions."
+            }
+        ],
+        description = "Recommended starter files with condition descriptions."
+    )
+    sample_data_path: Path = Field(
+        default     = Path("data/samples/wrf_sample.nc"),
+        description = "Local path where sample NetCDF file will be stored after extraction."
+    )
+    sample_data_url: str = Field(
+        default     = "https://huggingface.co/datasets/Jybbs/sfire-samples/resolve/main/samples.tar.gz",
+        description = "Hugging Face direct download URL for sample data tar.gz file."
+    )
+    sample_extract_dir: Path = Field(
+        default     = Path("data"),
+        description = "Directory where sample tar.gz will be extracted."
+    )
+    source: Literal["sample", "wrf-sfire", ""] = Field(
+        default     = "",
+        description = "Data source to download: 'sample' for quick start, 'wrf-sfire' for full dataset."
+    )
+    transfer_timeout: int = Field(
+        default     = 86400,
+        description = "Maximum seconds to wait for transfer completion (default: 24 hours)."
+    )
+    wrf_sfire_dir: Path = Field(
+        default     = Path("data/wrf-sfire"),
+        description = "Local directory for storing WRF-SFIRE dataset files from Globus."
+    )
+
+
+class GlobusSecrets(BaseSettings):
+    """
+    Secure storage for Globus OAuth2 tokens.
+    
+    Uses Pydantic's BaseSettings with secrets_dir for automatic persistence.
+    Each token field is stored as a separate file in the secrets directory,
+    with the filename matching the field name.
+    """
+    refresh_token: Optional[SecretStr] = Field(
+        default     = None,
+        description = "Long-lived token used to obtain new access tokens"
+    )
+    scope: Optional[str] = Field(
+        default     = None,
+        description = "Space-delimited OAuth2 scopes granted by this token"
+    )
+    secrets_path: Path = Field(
+        default     = SECRETS_DIR,
+        description = "Directory for storing secret files"
+    )
+    
+    model_config = SettingsConfigDict(
+        case_sensitive = False,
+        secrets_dir    = str(SECRETS_DIR) if SECRETS_DIR.exists() else None
+    )
+    
+    @computed_field
+    @property
+    def is_valid(self) -> bool:
+        """
+        Check if all required token fields are present.
+        """
+        return all(
+            getattr(self, field) is not None 
+            for field in ['refresh_token', 'scope']
+        )
+
+
 class MessagesModel(BaseModel, extra="forbid"):
     """
     Comprehensive message configuration for all CLI output.
@@ -279,6 +402,17 @@ class MessagesModel(BaseModel, extra="forbid"):
     monitoring_dynamics: str = Field(
         default     = "Monitoring thermal constraints and flock dynamics",
         description = "Message about training monitoring."
+    )
+    override_syntax_help: str = Field(
+        default = (
+            "# Override examples:\n"
+            "optimizer.learning_rate=0.001     # Learning rate\n"
+            "experience.batch_size=64          # Batch size\n"
+            "flock.agent_count=10              # Number of agents\n"
+            "hardware.precision=32-true        # Training precision\n"
+            "wandb.mode=offline                # W&B logging mode"
+        ),
+        description = "Help text for configuration override syntax."
     )
     ready_to_train: str = Field(
         default     = "Ready to train some thermal flocks? 🔥",
@@ -350,14 +484,95 @@ class MessagesModel(BaseModel, extra="forbid"):
         },
         description = "Wandb integration display messages."
     )
-    override_syntax_help: str = Field(
-        default = (
-            "# Override examples:\n"
-            "optimizer.learning_rate=0.001     # Learning rate\n"
-            "experience.batch_size=64          # Batch size\n"
-            "flock.agent_count=10              # Number of agents\n"
-            "hardware.precision=32-true        # Training precision\n"
-            "wandb.mode=offline                # W&B logging mode"
-        ),
-        description = "Help text for configuration override syntax."
+
+
+class PromptsModel(BaseModel, extra="forbid"):
+    """
+    Interactive prompt and dialog configuration.
+    
+    This model contains all settings for user interaction including error messages,
+    prompt templates, and questionary styling. It manages the conversational flow
+    of the CLI application.
+    """
+    config_not_found_error: str = Field(
+        default     = "Configuration '{config}' not found in available workloads",
+        description = "Error message template for missing configurations."
     )
+    confirm_override_prompt: str = Field(
+        default     = "Apply {count} override(s)?",
+        description = "Confirmation prompt for applying overrides."
+    )
+    override_prompt: str = Field(
+        default     = "Enter override value for {field_name}:",
+        description = "Prompt template for configuration overrides."
+    )
+    package_missing_error: str = Field(
+        default     = "Required package '{package}' is not installed",
+        description = "Error message template for missing packages."
+    )
+    python_version_error: str = Field(
+        default     = "Python {current} detected, but {required} or higher is required",
+        description = "Error message template for Python version mismatch."
+    )
+    questionary_style: list[tuple[str, str]] = Field(
+        default = [
+            ('question',    'fg:#ff6b6b bold'),
+            ('answer',      'fg:#4ecdc4 bold'),
+            ('pointer',     'fg:#ffe66d bold'),
+            ('highlighted', 'fg:#ff6b6b bold'),
+            ('selected',    'fg:#4ecdc4'),
+            ('separator',   'fg:#95e1d3'),
+            ('instruction', 'fg:#f38181'),
+            ('text',        'fg:#ffffff'),
+            ('disabled',    'fg:#808080 italic'),
+        ],
+        description = "Questionary prompt styling configuration."
+    )
+    workload_selection_prompt: str = Field(
+        default     = "Select a workload configuration:",
+        description = "Prompt shown when user needs to select a workload."
+    )
+
+
+class SystemModel(BaseModel, extra="forbid"):
+    """
+    System inspection and validation configuration.
+    
+    This model defines system requirements and validation rules used
+    by the SystemInspector to check environment compatibility.
+    """
+    cuda_preferred: bool = Field(
+        default     = True,
+        description = "Whether CUDA GPU acceleration is preferred for training."
+    )
+    dataset_validation: dict[str, float] = Field(
+        default = {
+            "min_size_gb"    : 0.1,
+            "max_size_gb"    : 10000.0,
+            "warning_size_gb": 100.0,
+        },
+        description = "Dataset size validation thresholds in gigabytes."
+    )
+    mujoco_min_version: str = Field(
+        default     = "2.3.0",
+        description = "Minimum MuJoCo version required for physics simulation."
+    )
+    python_min_version: tuple[int, int] = Field(
+        default     = (3, 9),
+        description = "Minimum Python version required as (major, minor) tuple."
+    )
+    required_packages: list[str] = Field(
+        default = [
+            "torch",
+            "pytorch_lightning",
+            "torchrl",
+            "mujoco",
+            "hydra-core",
+            "wandb",
+            "rich",
+            "typer",
+        ],
+        description = "List of required Python packages for system validation."
+    )
+
+
