--- src/config/cli/workloads/__init__.py
@@ -1,6 +1,0 @@
-"""
-CLI configuration workloads.
-
-This module contains the composed configuration that brings together all CLI
-schemas and factories into a single, instantiatable configuration object.
-"""
